+++ NEWS.md
@@ -1,3 +1,9 @@
+# sphereTessellation 1.1.0.9000
+
+The package does no longer depend on the 'randomcoloR' package. Instead, it 
+now depends on the 'colorsGen' and 'Polychrome' packages.
+
+
 # sphereTessellation 1.1.0
 
 - Now the package does not depend on the s2 package.